Rename package and module to be more descriptive
Can now be run with:
python scripts/configure_defaultSession \
-y scripts/configure_defaultSession/tests/fixtures/config.yml -o
