+++ scripts/configure_defaultSession/tests/fixtures/config.json
@@ -91,6 +91,34 @@
                 "configuration": "Lten_pop08_TD.bed-LinearArcDisplay"
               }
             ]
+          },
+          {
+            "id": "ther_default_Test_of_bedGraph_track",
+            "type": "QuantitativeTrack",
+            "configuration": "Lten_pop08_TD_for_wiggle.bed",
+            "minimized": false,
+            "displays": [
+              {
+                "id": "ther_default_Test_of_bedGraph_track_display",
+                "type": "LinearWiggleDisplay",
+                "heightPreConfig": 150,
+                "configuration": "Lten_pop08_TD_for_wiggle.bed-LinearWiggleDisplay"
+              }
+            ]
+          },
+          {
+            "id": "ther_default_Test_of_GFF_addTrack:false",
+            "type": "FeatureTrack",
+            "configuration": "test_gene_track.gff",
+            "minimized": false,
+            "displays": [
+              {
+                "id": "ther_default_Test_of_GFF_addTrack:false_display",
+                "type": "LinearBasicDisplay",
+                "heightPreConfig": 150,
+                "configuration": "test_gene_track.gff-LinearBasicDisplay"
+              }
+            ]
           }
         ]
       },
@@ -183,6 +211,58 @@
         {
           "type": "LinearArcDisplay",
           "displayId": "Lten_pop08_TD.bed-LinearArcDisplay"
+        }
+      ]
+    },
+    {
+      "type": "QuantitativeTrack",
+      "trackId": "Lten_pop08_TD_for_wiggle.bed",
+      "name": "Test of bedGraph track",
+      "assemblyNames": [
+        "Linum_tenue_thrum_v1"
+      ],
+      "adapter": {
+        "type": "BedGraphAdapter",
+        "bedGraphLocation": {
+          "uri": "Lten_pop08_TD_for_wiggle.bed.bgz"
+        },
+        "index": {
+          "location": {
+            "uri": "Lten_pop08_TD_for_wiggle.bed.bgz.csi"
+          },
+          "indexType": "CSI"
+        }
+      },
+      "displays": [
+        {
+          "type": "LinearWiggleDisplay",
+          "displayId": "Lten_pop08_TD_for_wiggle.bed-LinearWiggleDisplay"
+        }
+      ]
+    },
+    {
+      "type": "FeatureTrack",
+      "trackId": "test_gene_track.gff",
+      "name": "Test of GFF addTrack:false",
+      "assemblyNames": [
+        "Linum_tenue_thrum_v1"
+      ],
+      "adapter": {
+        "type": "Gff3TabixAdapter",
+        "gffGzLocation": {
+          "uri": "test_gene_track.gff.bgz"
+        },
+        "index": {
+          "location": {
+            "uri": "test_gene_track.gff.bgz.csi"
+          },
+          "indexType": "CSI"
+        }
+      },
+      "displays": [
+        {
+          "type": "LinearBasicDisplay",
+          "displayId": "test_gene_track.gff-LinearBasicDisplay"
         }
       ]
     }
